requirements.txt
streamlit>=1.31.0
Pillow>=10.2.0
urllib3>=2.0.7
openai>=1.12.0
